+++ app/helpers/agents_helper.rb
@@ -27,4 +27,14 @@ module AgentsHelper
     params["user_ids"] = user_ids if user_ids.present?
     params
   end
+
+  def display_meta_note(note)
+    meta = content_tag(:span, "le #{l(note.created_at.to_date)}", title: l(note.created_at))
+    meta += " par #{note.agent.full_name_and_service}"
+    content_tag(:span, meta, class: "font-italic")
+  end
+
+  def agents_to_sentence(agents)
+    agents.map(&:full_name_and_service).sort.to_sentence
+  end
 end
