+++ app/helpers/rdvs_helper.rb
@@ -16,32 +16,6 @@ module RdvsHelper
     "#{user.full_name} <> #{rdv.motif&.name}"
   end
 
-  def agents_to_sentence(rdv)
-    rdv.agents.map(&:full_name_and_service).sort.to_sentence
-  end
-
-  def users_to_links(rdv)
-    safe_join(rdv.users.order_by_last_name.map { user_to_link(_1) }, ", ")
-  end
-
-  def user_to_link(user)
-    if user.organisations.include?(current_organisation)
-      link_to user.full_name, admin_organisation_user_path(current_organisation, user)
-    else
-      "#{user.full_name} - l'usager a été supprimé"
-    end
-  end
-
-  def users_to_sentence(rdv)
-    return rdv.users.map(&:full_name).sort.to_sentence if current_agent
-
-    users = []
-    rdv.users.each do |user|
-      users << user if user == current_user || current_user.relatives.include?(user)
-    end
-    users.map(&:full_name).sort.to_sentence
-  end
-
   def rdv_status_tag(rdv)
     content_tag(:span, Rdv.human_enum_name(:status, rdv.status), class: "badge badge-info")
   end
@@ -63,6 +37,19 @@ module RdvsHelper
       content_tag(:span, "Annulé", class: "badge badge-warning")
     elsif rdv.starts_at.future?
       content_tag(:span, "À venir", class: "badge badge-info")
+    end
+  end
+
+  def stats_rdv_path(status)
+    case controller_name
+    when "stats"
+      if params[:agent_id].present?
+        admin_organisation_agent_rdvs_path(current_organisation, params[:agent_id], status: status, default_period: true)
+      else
+        admin_organisation_rdvs_path(current_organisation, status: status, default_period: true)
+      end
+    when "users", "relatives"
+      admin_organisation_user_rdvs_path(current_organisation, params[:id], status: status)
     end
   end
